Solved concurrency issues
No longer depend on TrackedChunk's finalize() method being called,
although if finalize() is called, it will add itself to ChunkManager's
cleanup queue. This will catch any changes made by the user to the Chunk
directly, if they forget to call closeAll() or if the reference is lost.

I will need to clean up the ChunkManager's code at some point, I've made
so many changes it's a bit disorganized. However, ugly as it is, it
seems to be working properly.